feat: add ability to init PCA separately from creation
There is a new async init() method on the PublicClientApplication class.
The code in this method used to be called directly from the constructor.
However, the constructor is not async so you couldn't know when the initialization was done,
and, if it threw errors, you couldn't catch them.
Since making the constructor async would be jank,
this async code is now extracted to the init() method.

To use, create the PCA instance with the new second "init" constructor parameter set to false:
`const pca = new PublicClientApplication(config, false)`.
Then call `await pca.init()` in a try/catch block.
It will throw an error if there is something wrong with your configuration.

The second "init" constructor parameter is optional and defaults to true,
so it will behave the same as before when not provided.